pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyca"
version = "0.1.0"
authors = [
    { name = "Vassilis Papadopoulos", email = "vassilis.physics@gmail.com" },
    { name = "Marco Tuccio", email = "marco.tuccio95@gmail.com" },
]
description = "PyCA, a small python package to facilitate implementing artificial life worlds with pytorch."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "easydict",
    "matplotlib",
    "numpy",
    "pygame-ce",
    "pyperlin",
    "showtens",
    "torch",
    "torchvision",
    "torchenhanced",
    "pygame-gui",
]

[project.optional-dependencies]
nca = [
    "tqdm",
    "wandb",
    "Pillow",
    "torchenhanced",
]

[tool.hatch.build.targets.wheel]
packages = ["pyca"]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }